🔵 move helper functions to assertion_helpers.go
